app-crypt/gnupg-pkcs11-scd: add 9999 ebuild
This project depends upon the old libassuan v2 due to a build bug
which got fixed upstream in the latest "master" branch, however
upstream releases are very far in between (last was in Jan 2022)
so we add the 9999 ebuild which works nicely with libassuan 3.0.

I've asked upstream to cut a new release containing the fix [1]
however I don't know how long that will take and we can still
have the same problems down the line, so it's useful to have
a 9999 git ebuild.

Link: [1] https://github.com/alonbl/gnupg-pkcs11-scd/issues/65
Signed-off-by: Adrian Ratiu <adrian.ratiu@collabora.com>