app-text/calibre: require minimum ffmpeg version
It simply fails to compile, with lower versions of ffmpeg. No revbump,
anyone successfully building doesn't need to recompile as it will be
slotted.

The alternative here is to add ifdef'ed support for ffmpeg 4 to
upstream. And upstream is unenthusiastic about this, for which I don't
blame him. Sorry to all the VLC users I guess... maybe you can finally
learn to love mpv?

Closes: https://bugs.gentoo.org/941926
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>