x11-drivers/nvidia-drivers: make egl-x11 X+wayland conditional
Upon closer look, this should only be for xwayland, so can skip
a few dependencies for pure X users. Pure wayland users (no
xwayland), also shouldn't need it.

Should likely package this separately like egl-wayland too, but
currently egl-x11 has no releases and little activity not sure
if we should be using it yet. Overall it'd be simpler to just use
the provided egl-gbm/wayland/x11 but this is about building from
source when we can.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>