profiles/thirdpartymirrors: add long-term fallback for apache
Apache has a mirror network, which only covers the most recent release
of any given package. They also have an additional site which does not
appear to be a CDN, and is throttled and can maybe ban you if you use it
too much. Unfortunately, it is also the ONLY way to actually get
historic releases of many packages.

Use it because it is the only way to guarantee actual availability. It
is now one of 7 possible mirrors -- even though practical
implementations of PMS randomly choose the mirror ordering, there's
still a great chance this will only be used after another mirror
provided any file available on the CDN.

In combination with GENTOO_MIRRORS this should ensure that users can
actually download software when needed...

... without running afoul of throttling. Maybe. Probably. A proper
tiering/priority system would need to be added to PMS if we wanted to
guarantee that, but the odds are favorable and users interacting
directly with upstream distfiles sites always run some form of potential
throttling risk.

Bug: https://bugs.gentoo.org/640600
Bug: https://bugs.gentoo.org/872182
Bug: https://bugs.gentoo.org/911774
Bug: https://bugs.gentoo.org/942151
Bug: https://bugs.gentoo.org/946318
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>