x11-drivers/nvidia-drivers: drop vuln 525.147.05-r1, 560.35.03-r1
Both branch are unsupported and so did not get security updates.
560 had short support due to being a New Feature Branch (NFB).

Users of ~testing 560.35.03-r1 are expected to downgrade to the
(newer) 550.127.05 version which is the next stable candidate if
no problems.

If for one reason or another the 550 branch was problematic for a user,
they may optionally want to opt-in the ~565.57.01 beta instead which is
not vulnerable (we do not keyword betas, see bug #941991 comment #1 --
but it can be manually accepted).

Users of 525.x are on their own, if *really* need that version and
cannot upgrade due to regressions then will have to keep it in a local
overlay. Alternatively the still supported 470.x may still be usable.

(there are still other vulnerable versions to drop but these are
awaiting stabilizations)

Bug: https://bugs.gentoo.org/942031
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>