x11-drivers/nvidia-drivers: keyword 565.57.0-r2 beta
Not particularly happy about this (esp. given this version is
meant to be used with egl-wayland-1.1.17 and this enables explicit
sync that may still cause some applications to crash), but with
560 being vulnerable and 550 not working so well for "some" users
on wayland (presumably those with hybrid graphics?), guess we need it
albeit this may solve issues for some and cause more for others...

If have problems, please downgrade to stable 550, can also downgrade
egl-wayland albeit that may hurt xwayland performance or cause visual
glitches with egl-x11.

For Xorg users, 565 or 550 should make close to no difference though,
may potentially have an impact on suspend/resume (good or bad).

Closes: https://bugs.gentoo.org/941991
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>