sys-boot/grub: support adding entries for UKIs using grub-mkconfig
This adds a uki USE flag supporting the installation of an extra grub.d
script. When grub-mkconfig is called it will the search for UKIs
matching the patterns "/boot/*EFI/Linux/*-*.efi" or
"/efi/EFI/Linux/*-*.efi", or those of the environment variable
$GRUB_UKI_GLOBS, if defined, and create boot entries for them.  05_uki
was chosen since it should be run immediately before or after 10_linux
and 05 was an multiple of 5 unused by other scripts.

Closes: https://bugs.gentoo.org/942201
Signed-off-by: Peter Levine <plevine457@gmail.com>