app-misc/binwalk: add 3.1.0 and update live ebuild
Original upstream was forked due to it being dead, supposedly.
Now it was active again, this new version switches to the original
upstream.
Since this new version is a completely rewrite in a different language a
slot has been used so users can play around with it while being
able to keep using the legacy version, if needed.

Signed-off-by: David Roman <davidroman96@gmail.com>
Closes: https://bugs.gentoo.org/942100
Signed-off-by: Sam James <sam@gentoo.org>