sci-mathematics/polymake: un-keyword for x86
Despite the fact that USE=lrs is masked for x86, pkgcheck is throwing a
fit about the fact that I've un-keyworded sci-libs/lrslib for x86 (it
doesn't even build on x86), and is turning the CI red. So to fix that,
I'm un-keywording polymake too.

I'll file a bug for x86@, but this is a pretty niche application and
probably nobody cares.

Signed-off-by: Michael Orlitzky <mjo@gentoo.org>