virtual/service-manager: new revision sans virtual/daemontools
Daemontools is DJBware that gets pulled in as a dependency of a few
other packages. It is also technically a service manager, but one that
is not seriously used today. This creates a problem: if users install a
package that pulls in daemontools, emerge --depclean now thinks that it
is safe to remove OpenRC, because this virtual is satisfied by
daemontools instead.

In some cases, technically correct is not the best kind of correct, and
this is one of those cases. Nobody is using daemontools as a service
manager, but a lot of users have complained about their actual service
manager getting depcleaned. It's time to drop this.

Meanwhile, sys-apps/baselayout-prefix was treecleaned in 2022, so I've
cleaned up that dependency as well.

Closes: https://bugs.gentoo.org/803878
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>