net-im/synapse: restrict dev-python/treq due to multipart issue
dev-python/treq-24.9.1-r1 depends on dev-python/multipart which has a
name collision with <dev-python/python-multipart-0.0.12-r100 in runtime.
Older treq revision bundles dev-python/multipart in order to avoid this
issue.

The issue is addressed starting with synapse-1.116.0-r4 revision for
>=dev-python/python-multipart-0.0.12-r100, but we need to restrict treq
for older versions until it will be possible to stabilize 1.116.0-r4.

Signed-off-by: Petr Vaněk <arkamar@gentoo.org>