dev-lang/rust{,-bin}: port to llvm-r1 and slot (-r100)
Port to llvm-r1 to gain the fancy new `llvm_slot_{x}` USE flags
which we use in the rust eclass to force a closer dependency on
specific LLVM slots.

Since Rust in Gentoo is only ever built against the one LLVM slot
we are able to simplify the LLVM logic in the Rust ebuilds and use
our knowledge of the LLVM -> Rust version mapping to enable
slots for dev-lang/rust in a usable manner.

Since Rust is now slotted and not managed entirely by eselect-rust
each slot needs to be added to LDPATH.

We also drop the `profiler` USE as it's inconsistent with what
upstream ship by default and casues issues where other package's
build systems assume that it exists.

This commit also introduces dev-lang/rust-common which handles bash
completions for slotted Rust.

dev-lang/rust-bin:

`LLVM_OPTIONAL` is used as we don't need to consume LLVM,
we really just want the free IUSE.

Closes: https://bugs.gentoo.org/941146
Signed-off-by: Matt Jolly <kangie@gentoo.org>