toolchain.eclass: Improve the search for libcc1 artifacts
When building cross GCC toolchain, we need to move the libcc1* artifacts
to a separate target dir. That's handled in our eclass, not by the GCC's
build system.

That mechanism was relying on these artifacts being installed in
`/usr/$(get_libdir)`. However, when such cross environment is being
created on an LLVM-based system (any LLVM stage3/profile), it ends up
in `/usr/lib`, even in cases when `$(get_libdir)` would evaluate to
`lib64`.

To solve that, check both possible locations.

Signed-off-by: Michal Rostecki <vadorovsky@protonmail.com>