media-libs/opencv: refactor tests. split up skip logic per module
The prior logic expanded CMAKE_SKIP_TESTS for each tests. Filtering testcases
that do exist for that test.
Split that up to be set per test.

Do not run the testphase for abi_x86_32 as they are never build.

Skip CUDA related tests when we can not access the render devices. Usually due
to the build user not being in the video group.

Signed-off-by: Paul Zander <negril.nx+gentoo@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>