app-emulation/qemu-guest-agent: Check for VIRTIO_CONSOLE kernel config
To control guests from inside qemu-guest-agent is used. By
default, it listens on a virtio-serial device
(/dev/virtio-ports/org.qemu.guest_agent.0, see [1]) which is available iff
CONFIG_VIRTIO_CONSOLE kernel config is enabled. Due to a bug in
qemu-guest-agent where it daemonizes itself before opening the
device, init scripts are unable to tell whether startup was
successful. Check for the kernel config knob and hope that users
will notice it while emerging the package. There is a fix for the
bug [2] but it's contained in 10.0.0 release.

1: https://gitlab.com/qemu-project/qemu/-/blob/88f72048d2f5835a1b9eaba690c7861393aef283/qga/main.c#L46
2: https://gitlab.com/qemu-project/qemu/-/commit/c6f5dd7ac8ef62dcdec4cdeda1467c658161afff
Closes: https://bugs.gentoo.org/810628
Signed-off-by: Michal Privoznik <michal.privoznik@gmail.com>
Part-of: https://github.com/gentoo/gentoo/pull/43567
Signed-off-by: Sam James <sam@gentoo.org>