x11-plugins/wmbatteries: fix modern C porting in configure
Regenerate the configure script to take advantage of updates to core
autoconf macros so that system probes function correctly.

The software doesn't actually build with Modern C. That is tracked by a
different bug. It's an open question whether this package makes sense to
package anyway -- especially given it is only even *keyworded* for
x86...

Closes: https://bugs.gentoo.org/942786
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>