sys-libs/glibc: Enforce GCC also for cross builds
Before this change, overriding clang configuration with GCC was done
only for native builds. Doing the same for cross builds is a necessary
step to make LLVM+glibc cross environments possible to bootstrap. In
case of the full cross builds (not only headers, but the full libx), the
target-specific wrappers need to be used.

Signed-off-by: Michal Rostecki <vadorovsky@protonmail.com>