rust.eclass: dependency strings: dev-lang/rust-bin first
To better match the behaviour of the legacy virtual/rust package,
update generated dependency strings so that dev-lang/rust-bin is
output before dev-lang/rust.

In the longer term it is expected that this will be reverted
to the original order (prefer dev-lang/rust) as users that
particularly care about this also have the option of selecting
an appropriate binpkg.

When selecting an installed Rust to use the non-bin package is
still preferred over the -bin if both are installed.

Signed-off-by: Matt Jolly <kangie@gentoo.org>