www-client/firefox: LTO Updates
Force LTO with USE=pgo

LTO now comes from CFLAGS, meaning that the REQUIRED_USE
previously enforcing this was removed in favour of
`eerror`. Instead we will force LTO even if it's not detected
from CFLAGS if PGO is in use.

Drop USE=lto on 128 to bring it into line with 132.

Signed-off-by: Matt Jolly <kangie@gentoo.org>