cargo.eclass: check for and quote `CARGO`
A quick sanity check that `CARGO` is set in appropriate cargo_src_*
phase functions will make it easier for users to catch occurrences
where pkg_setup does not call rust_pkg_setup.

Also inherit llvm-r1 if required and let it die instead of us if
the environment is not properly setup.

Closes: https://bugs.gentoo.org/943147
Signed-off-by: Matt Jolly <kangie@gentoo.org>