dev-python/pip: fix minimum bounds for packaging dep
We were waiting on an update to packaging before bumping pip (as it
imports new API from packaging 24.2), but having gotten that update we
didn't actually depend on it.

This causes pip to fail to build, as packaging is used at build time in
order to create shell completions (and obviously for testing as well)
and nothing requires packaging to update at all, let alone requires it
to come first in the build queue.

Closes: https://bugs.gentoo.org/943156
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>