www-client/firefox: fix building with USE=pgo and no -flto
We automatically enable lto in this case. But we also run some
conditionals too early, only in the `tc-is-lto` check, that do not get
run under USE=pgo. Specifically, we filter some -Werror=* due to
reasons.

This then means that people setting these -Werrors globally, but only
selecting certain packages to build with -flto, cannot build firefox.

(The fact that these -Werrors are disabled is another topic entirely.)

Fixes: 01ba4dc61965ef7658a24728841c04c9a1ce4871
Fixes: bcec88143087b45ba26f219efa76021e5acac391
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>