profiles/base: mask www-client/chromium[system-icu]
Chromium fails to build with ICU 76.1.

While we could work around this in the ebuild, we already have machinery
in place that enables us to toggle the use of the bundled library;
we'll mask the USE for a little while until upstream have sorted it out.

Given how fast Chromium moves this should not stick around for too long.

Bug: https://bugs.gentoo.org/943216
Signed-off-by: Matt Jolly <kangie@gentoo.org>