dev-lang/rust: use rust.eclass for USE=system-bootstrap
Take advantage of rust.eclass for dependency generation and appropriate
bootstrap slot selection. Between the PATH manipulation that the eclass
performs and the exported `RUSTC` we should see significantly more
consistent bootstraps.

Without this change `eselect-rust` would determine which `rustc` to run,
resulting in build failures if an incorrect Rust was selected (which is
nearly always in the age of slotted Rust).

This is also one step towards removing USE=system-bootstrap entirely;
now that we can reliably (and trivially) select appropriate slots for
both dev-lang/rust and dev-lang/rust-bin we'll shortly be able to
avoid downloading bootstrap bins at all.

Bug: https://bugs.gentoo.org/943144
Closes: https://bugs.gentoo.org/943145
Signed-off-by: Matt Jolly <kangie@gentoo.org>