dev-python/PyQt6: add 6.8.0_pre2410211537
This may possibly not get released until (after) Qt6.8.1, and
would rather not mask 6.8.1 as well. So add this now (masked)
for testing, will unmask alongside adding 6.8.1 if still no
proper release.

Note have not tested with Qt6.8.0, but rather with 6.8.1 branch
using 6.8.9999. Odds are will not unmask Qt6.8.0 anyway.

This is supposed to need (unreleased) >=pyqt-builder-1.17 and
>=sip-6.9 per project.toml but still "seems" fine with the old
versions at the moment. May "possibly" need newer PyQt6-sip as
well at runtime, but (at least) qutebrowser tests been passing
without it. So avoiding bumping to a bunch of extra snapshots.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>