dev-lang/luajit: add 2.1.9999999999
This addresses the -9999 request in #843773.

Currently, upstream pushes all changes to the v2.1 branch.

Per http://luajit.org/status.html,

> The old git master branch is phased out and stays pinned
> to the v2.0 branch. Please follow the versioned branches
> instead.

I think that tracking v2.1 is in the spirit of the original
request; most people interested in this ebuild will want to
have the latest rolling commit of v2.1, and will not want to
suddenly be switched to a new incompatible version of
luajit. To be consistent, I've versioned it 2.1.9999999999
so we can create a 2.2.9999999999 or 3.0.9999999999 if the
time comes. I've included enough nines that we should be
good for the next hundred or so years ;)

Bug: https://bugs.gentoo.org/843773
Signed-off-by: Artemis Everfree <artemis@artemis.sh>
Closes: https://github.com/gentoo/gentoo/pull/39293
Signed-off-by: Sam James <sam@gentoo.org>