dev-scheme/guile: fix 32-bit patch
As Stefan reports on the forums [0], regular Make conditions don't
work as expected in Automake. We'd end up having GUILE_OPTIMIZATIONS
set but Automake not realising it, so it'd use the default -O2 below,
which made stage0 far slower than it needed to be.

Fix that so that -O1 is always used when it should be, and the special
-O1 ... options are needed for 32-bit BE, as originally intended. -O2
is now never used for stage0.

This fixes slow guile build times on platforms with prebuilt/.

(I will note, however, that ultimately, we may end up with slow build
times again if we either delete the prebuilt/ stuff entirely or make it
optional - which is indeed how I didn't notice it myself at least.)

[0] https://forums.gentoo.org/viewtopic-p-8845997.html#8845997

Bug: https://bugs.gentoo.org/940650
Signed-off-by: Sam James <sam@gentoo.org>