dev-lang/luajit: fix LJLIBD not relative to PREFIX
luajit contains a hardcoded path (LJLIBD) where it searches
for some lua libraries it installs. This patch was
previously defined relative to PREFIX. Upstream commit
ae4735f [1] changes them to be defined relative to DPREFIX,
which is in turn relative to DESTDIR. I believe this is a
bug, which I've reported upstream [2].

This commit patches the Makefile to define LJLIBD relative
to PREFIX again.

1: https://github.com/LuaJIT/LuaJIT/commit/ae4735f621d89d84758769b76432d2319dda9827
2: https://github.com/LuaJIT/LuaJIT/issues/1303
Closes: https://bugs.gentoo.org/943395
Closes: https://bugs.gentoo.org/943346
Signed-off-by: Artemis Everfree <artemis@artemis.sh>
Closes: https://github.com/gentoo/gentoo/pull/39311
Signed-off-by: Sam James <sam@gentoo.org>