dev-db/postgresql: build w/ -std=gnu17
Upcoming GCC 15 defaults to -std=gnu23 which pg fails to build with. Go
back to the default -std=gnu17 (i.e. a noop for older GCCs) until the
upstream report I've sent is fixed.

Sent that to pgsql-bugs but it's still in the moderation queue for now,
I'll update with the link once it's through.

Signed-off-by: Sam James <sam@gentoo.org>