sys-fs/progsreiserfs: mark as LTO-unsafe on musl
It fails with linker errors due to including a code copy of a manky old
libintl.

Happens because for some inexplicable reason, packages which genuinely
need the full power and flexibility of GNU libintl, can only rely on
dev-libs/libintl on *non* linux operating systems, even though it was
very much intended upstream to be useful and usable on musl systems.

glibc, obviously, works fine.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>