dev-lang/rust: libexec-${PV} symlink requires USE=rust-analyzer
Until 1.72.0 `cargo-credential-1password` was installed in
`/usr/lib/rust/${PV}/libexec`, and we were able to unconditionally
symlink that for eselect-rust to use.

This is no longer the case and most of the time users end up
with a broken symlink for each rust slot in /usr/lib/rust.

Gate the symlink creation with the appropriate USE and
conditionally add `/usr/lib/rust/libexec` to the
eselect-rust provider config.

Closes: https://bugs.gentoo.org/915403
Signed-off-by: Matt Jolly <kangie@gentoo.org>