x11-drivers/nvidia-drivers: bump KERNEL_MAX to 6.12 where possible
Don't really recommend 6.12 yet, but it *should* mostly work.

There *may* be issues for which it is unclear if they were
limited to older 6.12-rc1 or so and not the actual 6.12.
It's possible that some issues only surface with specific
kernel configs and so they're hard to pick up.

Normally don't do patches, but in this case it compiles even
though it's going to be semi-broken without it at runtime.
FOP_UNSIGNED_OFFSET is unset for <6.12 so this should have
no impact for older kernels.

0/vulkan fails to build, and 0/470 was already broken with 6.10.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>