dev-qt/qtmultimedia: skip more tests failing w/ sandbox/accel (qt6)
A change in tst_qmediaplayer_gstreamer causes it to attempt to use
real pulseaudio which doesn't work here. Might add that upstream
recently starting skiping all gstreamer tests (QTBUG-129469) in
their CI while recognizing that they're flaky.

And tst_qmediarecorderbackend seems to be trying to use ffmpeg+nvenc
for me which does not end well... May possibly pass depending on
USE on ffmpeg but rather not have to worry about this.

On a side-note, this was the last known (non-)issue with Qt 6.8.1
branch and it should be ready for release.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>