sys-devel/gcc: backport C23 fix to 15.0.0_pre20241117
Backport a -std=gnu23/-std=c23 fix to 15.0.0_pre20241117 which prevented
coersion from _Bool to NULL. Note that this is still dodgy code and some
are proposing this be banned in a future C version (maybe even C2y), but
it shouldn't be rejected in this version at least.

I've added Closes tags for all bugs which, at a glance, look related,
but I haven't re-tested them all and one or two might need reopening.

Bug: https://gcc.gnu.org/PR112556
Closes: https://bugs.gentoo.org/943862
Closes: https://bugs.gentoo.org/943861
Closes: https://bugs.gentoo.org/943843
Closes: https://bugs.gentoo.org/943811
Closes: https://bugs.gentoo.org/943809
Closes: https://bugs.gentoo.org/943753
Closes: https://bugs.gentoo.org/943715
Closes: https://bugs.gentoo.org/943704
Closes: https://bugs.gentoo.org/942958
Signed-off-by: Sam James <sam@gentoo.org>