java-pkg-simple.eclass: support multi-release according to JEP 238
This commit adds basic support for building multi-release jar files
according to JEP 238. A multi-release jar file has release-specific
classes in directories under META-INF/versions/ and its MANIFEST.MF
contains a line with 'Multi-Release: true'.

To do so, it adds JAVA_RELEASE_SRC_DIRS as a new eclass variable which
is also used as the condition to trigger the new functionality. A new
local variable 'multi_release' is added to the packaging section (the
part using the 'jar -create' command). Only when JAVA_RELEASE_SRC_DIRS
is set, additional actions take place:

- Compilation (those are the parts with 'ejavac') will additionally loop
  over the release-specific directories listed in JAVA_RELEASE_SRC_DIRS
  and compile the release-specific classes into corresponding directories
  under target/versions/.

- Packaging (the part using the 'jar -create' command) will add the
  details to the 'multi_release' variable so that the release-specific
  directories under target/versions/ can be packaged into the jar file.

There are packages where module-info.java is not provided in the sources
but needs to be generated by the build system. We use the built-in jdeps
function with the --generate-module-info option which became available with
Java 11.

It generates the module-info.java file based on an intermediate jar file
which needs to be named as "${JAVA_MODULE_NAME}.jar" and places it in
the "${JAVA_MODULES_DIR}/${JAVA_MODULE_NAME}/versions/9" directory.

For this purpose we add two new eclass variables:
- JAVA_MODULE_NAME
- JAVA_MODULES_DIR

When both JAVA_MODULE_NAME and JAVA_MODULES_DIR are defined in the
ebuild we
- compile the sources still without module-info
- package them as an intermediate {JAVA_MODULE_NAME}.jar
- let java-pkg-simple_generate-module-info generate the module-info
- compile module-info.java with the --patch-module option
- package the final jar file with the module-info.class in a directory
  specific for --release 9.

Bug: https://bugs.gentoo.org/900433
Signed-off-by: Volkmar W. Pogatzki <gentoo@pogatzki.net>