x11-drivers/nvidia-drivers: lower KERNEL_MAX to 6.6 for 0/470
NVIDIA's support seemingly has ended, as such this will never
go up anymore (unless we patch it, which there is no plans to
support). It does not make sense to list a non-LTS branch that's
not even in the tree anymore there.

Also add a warning similar to 390.x, the date is the same given
both 6.1.x and 6.6.x kernels are going away at same time.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>