profiles: mask nvidia-drivers:0/470 EOL branch
Much like we already for (still-in-tree) 0/390, mask message
is mostly an updated copy/paste. I was under the impression
EOL wasn't already reached but no... and there are already
known security issues wrt bug #942031.

As noted in the mask message, users are free to unmask for
the time being but should be aware that it could be removed
anytime (late 2026, or early 2027 are only a tentative dates).

So, before then should either switch to nouveau, try to get
new hardware, or be prepared to be on your own to keep using
these drivers (much like formerly removed 340.x), some overlays
may possibly support these but well.

Bug: https://bugs.gentoo.org/942031
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>