x11-drivers/nvidia-drivers: add 550.135
incl. proper upstream testing & fix for kernel 6.12, would recommend
this version over the 565.57.01 beta if want to use 6.12 (albeit beta
should be ok with the patch too, just no guarantees that it isn't
missing something).

Also newly needs a minor workaround given NVIDIA newly expects ARCH
to be set to x86_64 while tc-arch-kernel sets x86 (with the rationale
that it makes no difference for the kernel). There is however little
reason to "assume" that the kernel (or modules) sees it the same way
and maybe tc-arch-kernel should pass x86_64... but for now change it
here if USE=amd64 given unsure if want to look into other potential
implications of changing this.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>