dev-python/twisted: add more IDEPEND for regen-cache
One additional IDEPEND is needed that was missed in the previous
commit on this issue (d74d1d24e34cfd796a297695d05fec023a783701).
twisted-regen-cache should run without error now that
typing-extensions is pulled in. This was tested from an image
with nothing else installed so should finally have everything.

One note is that incremental is also technically required at
install time and may be missing, however there are no real cases
where this is an issue because it is in both BDEPEND and RDEPEND
and emerging with --nodeps does not install IDEPEND so the fact
that install would fail in that case is expected.

Signed-off-by: Tom Gillespie <tgbugs@gmail.com>