dev-python/meson-python: BDEPEND on RDEPEND
Fix issue where meson-python fails to build if
pyproject-metadata is not installed at build time.
This is needed e.g. when building a binpkg. The
full RDEPEND is added because technically both
meson and ninja are required at build time and
are present in the base stage3 image, however if
there are cases where they are not present then
then emerge would fail.

Signed-off-by: Tom Gillespie <tgbugs@gmail.com>
Closes: https://github.com/gentoo/gentoo/pull/39403
Signed-off-by: Michał Górny <mgorny@gentoo.org>