x11-drivers/nvidia-drivers: add (our own) PREEMPT_RT check for 6.12
Would make it a hard failure (no ~) given a NVIDIA sanity check already
causes the build to fail, but there is a variable to override that.

Doesn't hurt to do our own standard check, mention the variable, and
warn that using it is unsupported given the increasing amount of users
jumping on PREEMPT_RT (that for most they likely do not need nor
understand for who it's useful for, and notably that it will also hurt
performance).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>