app-emulation/wine-vanilla: add 9.22
Have not tested with wayland myself (will give it a try when it's
time to stabilize wine-vanilla-10.0), but upstream changelog states:

    - Wayland driver enabled in default configuration.

I "assume" this means that it will now use wayland without requiring
custom configuration, albeit "believe" it'll still try x11 first
given `default_driver[] = L"mac,x11,wayland";` ordering so it'll
"likely" use xwayland if USE=X and DISPLAY is set and likely does not
warrant an ebuild warning for what would otherwise be a sudden change.

Either way, this "likely" fixes bug #939792 which had USE="-X wayland".
wine-staging-9999 (that the bug listed) may or may not already have
the changes (if not, will likely do when it rebases for 9.22) so going
ahead and closing it.

For anyone reading this, do keep in mind that wayland support is still
kind of WIP.

Closes: https://bugs.gentoo.org/939792
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>