dev-lang/rust: Drop patch that stops setting CMAKE_ASM_COMPILER for bundled LLVM
It is necessary to set CMAKE_ASM_COMPILER when cross-compiling, otherwise the
ASM compiler gets detected with the wrong tuple. The patch was added because the
build previously complained that CMAKE_ASM_COMPILER was unused, but that no
longer seems to be the case, at least since 1.82. This was tested with a cross
and a native build.

Bug: https://bugs.gentoo.org/933382
Signed-off-by: James Le Cuirot <chewi@gentoo.org>