profiles: drop obsolete mask for Qt 6.8.0 & friends
No longer needed now that it'll be overshadowed by 6.8.1, and
this pyqt6 snapshot needs to be unmasked to be used with 6.8.1.

Probably won't keep 6.8.0 in-tree for very long given it was
masked the whole time and is not a stable candidate, but just
a bit in case we want to compare regressions with 6.8.0 rather
than 6.7.3.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>