x11-drivers/nvidia-drivers: pass -std=gnu17 for gcc15 for now
Seems to be quite a few issues and may not be trivial to fix, upstream
already has some sanity checks to pickup that conftests failed and it
should be fixed properly in time.

Using KERNEL_CC to ensure it's used everywhere for modules, esp.
conftest.sh. Non-modules parts seems fine with c23.

For 390/470, just add it to the list of other permanent (ugly)
workarounds and update's 470 comment given it's no longer supported.
Due for removal from the tree in roughly 2 years or less and so just
need to hold on until then.

Closes: https://bugs.gentoo.org/944092
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>