dev-util/ruff: add 0.8.0
Needs one backport to fix a test regression when not building with
debug.

I've also added an experimental dedup thing. One of the major issues
with rust is that cargo locks and requires all crates to be downloaded
offline, even if you're not using most of them because the crate
dependencies are Windows-specific. I haven't solved that, but I did
manage to remove about 40 of them by removing some workspace members
from the ruff monorepo that aren't actually part of the ruff binary in
any configuration (which means their dependencies don't need to be
locked, either). This has to be done before running pycargoebuild *and*
also done in the ebuild itself.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>