app-emulation/wine-proton: allow using compiler-rt[atomic-builtins]
Passing -latomic is no longer needed when using the new compiler-rt
w/ USE=atomic-builtins, and lets gcc be optional on llvm profiles.

To recap:
* gcc: don't need -latomic
* clang (w/o compiler-rt): don't need -latomic
* clang + compiler-rt[atomic-builtins]: don't need -latomic
* clang + compiler-rt[-atomic-builtins]: needs -latomic

So, if compiler-rt always provides atomics in the future (rather
than optional) all workarounds could be dropped.

The || ( ) is a bit "best effort" given dependencies don't know
what is actually going to be used.

(not worth a revbump, if really want to depclean gcc please
re-emerge manually)

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>