media-gfx/blender: Remove the experimental option
The "experimental" flag is only intended to be used in development to
try out new features. It is untested in regular releases and can cause
the .blend files to not be compatible with other Blender versions.

This is turned on/off by the Blender developers in the default cmake
settings. Forcing it to something else than what is the default will
in most cases lead to issues.

Signed-off-by: Sebastian Parborg <darkdefende@gmail.com>