app-emulation/virt-manager: Fix shebang of installed files
Previously, app-emulation/virt-manager was installed using
distutils which automatically fixed shebang on installed python
scripts. Then virt-manager switched to meson, and our ebuilds
simply started inheriting meson instead of distutils-r1. But this
alone is not enough as shebangs are now broken. Fortunately,
there's python_fix_shebang which can be called to fix shebangs
and in ideal world this function would be called in src_prepare()
but this is not the case. virt-manager and other files are
generated at compile time with also too generic shebang [1].
Therefore, let's do that at the end of src_install().

1: https://github.com/virt-manager/virt-manager/blob/main/scripts/make_bin_wrapper.py
Fixes: c5ca13d818262729e691e4d23487cb427b50665d
Closes: https://bugs.gentoo.org/945329
Signed-off-by: Michal Privoznik <michal.privoznik@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>