dev-lang/rust: update non-slotted rust library tidyup
We would previously `die` if there was only one file found matching
libraries in ${old_rust_libs} after the fingerprint was stripped.

It turns out that it's not uncommon for fingerprints to match, and
it is entirely possible for one (or most!) libraries to actually
be overridden.

Keep the existing duplicate handling but log the `expected 2 but found
1` case just in case we need to track this down.

Since all -r100 ebuild are having `pkg_postinst` updated, be consistent
about `optfeature` usage.

Signed-off-by: Matt Jolly <kangie@gentoo.org>