llvm-utils.eclass: Add `llvm_cmake_usex_musl` helper
Checking only for `elibc_musl` USE flag is not sufficient when building
a musl cross environment on a glibc host. In such case, the host profile
always enforces `-elibc_musl` and effectively breaks the build of LLVM
toolchain (llvm-libunwind and libcxx*).

To cover that case, provide the helper which echos `ON` under two
ciricumstances:

* `CTARGET` is `*-musl*`.
* `elibc_musl` USE flag is set.

Signed-off-by: Michal Rostecki <vadorovsky@protonmail.com>