sys-libs/libcxx: Support cross build
Inherit the `crossdev` eclass and use a cross sysroot when a custom
target is specified. That allows to bootstrap libcxx with crossdev.

Extend the compiler test with the case when the (cross) toolchain has
a runtime, but doesn't have stdlib yet (because it's about to be built).
In such case, a build with default arguments fails, but a build with
`-nostdlib` succeeds. libc++ doesn't need stdlib to build, so that's
fine.

Signed-off-by: Michal Rostecki <vadorovsky@protonmail.com>
Signed-off-by: Michał Górny <mgorny@gentoo.org>