app-antivirus/clamav: add rust_pkg_setup
This does not cause any changes to package behaviour;
we don't depend on LLVM or have any version restrictions on
Rust so it doesn't matter which slot was picked - it's probably
the `eselect`ed one anyway.

There's no need to force a revbump on users who have already
built the package, however we can provide consistent selection
of dev-lang/rust via the eclass and get some output on the
selected slot as part of the build log if we fix this now.

Signed-off-by: Matt Jolly <kangie@gentoo.org>