sys-auth/sssd: add 2.9.6
The "sudo" USE flag was removed as it did not change dependencies.
Python support optional, so merged python support in.
Subid support is in all shadow packages, so merged support in.
man page are unconditionally built, in line with Gentoo standards.
The corresponding USE flags were dropped. The "passkey" "USE flag
was added and FIDO2 support is available.

Signed-off-by: Christopher Byrne <salah.coronya@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>