dev-python/pyqt6: update sip+pyqt-builder min reqs
Old versions still "worked" but were technically incorrect missing
a few new things for Qt 6.8 (pyproject.toml also requests these
versions), but given new sip+pyqt-builder weren't released yet we
stuck to the old for the snapshots.

This will likely become enforced when PyQt6 gets a proper release.

Not that it was a big issue so shouldn't need to revbump.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>